bringing gamepad input a bit more in. help me with testing please - i can somewhat fake gamepad input with some utility but it's not too handy
git-svn-id: svn+ssh://svn.code.sf.net/p/supertuxkart/code/main/branches/irrlicht@3310 178a84e3-b1eb-0310-8ba1-8eac791a3b58

+++ src/tracks/track.cpp
@@ -936,7 +936,7 @@ void Track::loadTrack(const std::string &filename)
         if(m_sky_textures.size()!=6)
         {
             fprintf(stderr, "A skybox needs 6 textures, but %d are specified\n",
-                    m_sky_textures.size());
+                    (int)m_sky_textures.size());
             fprintf(stderr, "in '%s'.\n", filename.c_str());
 
         }
